backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# AI & API
openai
requests
python-dotenv

# Data & Logic
pydantic
pandas